[file] Fix relative symlinks to directories on Windows
Windows has two concepts of directory symlinks: actual soft symlinks,
added in Visa, and directory junction points.  Junction points cannot
support relative symlinks, but that was what we were using if we
detected that the target was a directory.  This resulted in all sorts of
strange behavior, such as calling `symlink("foo", link)` creating a
junction point pointing to the absolute path `joinpath(pwd(), "foo")` if
there was a directory named `foo` in the current directory, and a
relative symlink to a file `"foo"` otherwise.

This PR cleans this up significantly, allowing for the creation of
realtive directory symlinks on Windows (assuming you have Administrator
privileges, of course) and even allowing users to force creation of a
directory symlink even if we can't auto-detect that the target should be
a directory.  My best advice is to create symlinks only after the files
themselves you want to point to already exist, but let's go ahead and
be flexible since it doesn't cost us much.